+++ guniflask/templates/project/{{project_name}}/__init__.py
@@ -1,5 +1,3 @@
 # coding=utf-8
 
 from .app import create_app, db, config
-
-__all__ = ['create_app', 'db', 'config']