+++ guniflask/utils/model.py
@@ -63,3 +63,9 @@ def _get_ignore_set(ignore):
     if isinstance(ignore, str):
         ignore = [i.strip() for i in ignore.split(',')]
     return set(ignore if ignore else [])
+
+
+def inject_sqlalchemy_model(model_cls):
+    model_cls.to_dict = lambda self, **kwargs: model_to_dict(self, **kwargs)
+    model_cls.from_dict = classmethod(lambda cls, dict_obj, **kwargs: dict_to_model(dict_obj, model_cls=cls, **kwargs))
+    model_cls.update_by_dict = lambda self, dict_obj, **kwargs: update_model_by_dict(self, dict_obj, **kwargs)