+++ docs/changelog.rst
@@ -13,8 +13,9 @@ Change Log
 - 在 ``manage debug/start`` 命令中可以通过 ``-p`` 选项设置 ``active_profiles`` 对应的环境变量
 - ``manage initdb`` 会自动加载foo模块及子模块中所有声明的 ``db.Model``
 - Flask app的name更正为项目的名称
-- 项目配置相关文件统一命名为app，包括 conf/app.py , conf/app-env.sh , bin/app-config.sh
+- 脚本文件命名更新: conf/foo-env.sh -> conf/app-env.sh , bin/foo-config.sh -> bin/app-config.sh
 - 配置项由原来的 ``config.settings`` 改为直接通过 ``settings`` 对象获取
+- 移除 conf/wsgi.py
 
 0.2.5 (2018-09-18)
 ------------------