+++ docs/changelog.rst
@@ -14,7 +14,7 @@ Changelog
 - ``manage initdb`` 会自动加载foo模块及子模块中所有声明的 ``db.Model``
 - Flask app的name更正为项目的名称
 - 脚本文件命名更新: conf/foo-env.sh -> conf/app-env.sh , bin/foo-config.sh -> bin/app-config.sh
-- 配置项由原来的 ``config.settings`` 改为直接通过 ``settings`` 对象获取
+- 提供 ``settings`` 对象获取配置
 - 移除 conf/wsgi.py
 - 提供基于JWT的用户权限认证方案